replicators: Allow mixed replication offsets
If at startup we detect that we've successfully snapshotted some
tables (aka they have a non-None replication offset) but not others,
we'd *like* to be able to re-snapshot only the ones that haven't
successfully been snapshotted yet. The issue is that then we'd have some
tables that have correct data as of one replication offset (the ones
that were earlier successfully snapshotted), but other tables that have
correct data as of a *later* replication offset (the ones we just
re-snapshotted). If we get in this situation, we can work around this by
starting replication at the *older* replication offset, and then
discarding replication events that we receive for either individual
tables or the DDL that are *before* the recorded replication offset we
have for that object.

In the case where we've fully snapshot, though, we can do what we used
to do and just start at the maximum of all replication offsets, since we
know that we've fully replicated all events up to that point.

Under the hood, we're now querying the `replication_offsets()` RPC
endpoint rather than the singular `replication_offset()` endpoint, and
storing the set of all replication offsets as a field of the
replicators::NoriaAdapter

Refs: ENG-854
Change-Id: Ia118d9cd538a36ed1546e855a84d1e63bab753fb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1258
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>