server: Clarify naming for schema replication offset
To be consistent with the `schema` field of the `ReplicationOffsets`
struct, rename a lot of the controller-global replication offset
variables, fields, and endpoint from `set_replication_offset` to
`set_schema_replication_offset`. This is slightly more verbose, but much
clearer especially since we're more actively mixing up schema and
non-schema replication offsets.

Change-Id: I919e119c0a576ab11f48836035863574ece55f07
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1259
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>