replicators: Set schema offset before snapshotting tables
Since we now gracefully handle the case where we have some offsets
present but not others, we can safely mark that we've snapshotted the
schema before we fully snapshot the tables. As of this commit, if any
tables fail to snapshot we'll just re-run the entire snapshot operation,
but later we can optimize this by skipping the schema snapshot if we
already have a schema replication offset.

Refs: ENG-854
Change-Id: I7bef25e88096b526bc28b1c56029abe238e77615
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1261
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>