fault tolerance: Pass view name with view queries.
Views are used to query specific readers at workers keyed by their
NodeIndex, the graph node identifier, and shard. The worker maps
(NodeIndex, shard) to the readers upon receiving a query. However,
upon leader failure, the NodeIndex of each node can change. This
assignment may map a different reader to a different reader's old
NodeIndex. When this occurs we panic in the worker.

The panic was neccessary as Views can be cached and querying a
different reader than the intended one can return incorrect results
to the user.

To ensure that we always query the correct reader, we add the name of
the view to the reader map. Each reader is now keyed by:
 (NodeIndex, view name, shard)
As a result, we no longer need to panic if the reader already exists
as any cached view will still be querying the correct reader.

A better solution would be to ensure that NodeIndex values are stable
across leader failures. This would be encompassed in a solution that
doesn't replay every single recipe [ENG-838] to recreate the graph
on leader failure.

Change-Id: I53407470141222403aa203f66ec8bcf56535721c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1281
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>