replicators: Don't snapshot already-snapshotted tables
In the beginning of the replicator process during the initial snapshot
of all tables, skip snapshotting the schema if we already have a
replication offset for it, and skip snapshotting any tables for which we
already have a replication offset. This in combination with past changes
gets us table-level-granular snapshot reliability up to but not
including the point of table compaction.

Refs: ENG-854
Change-Id: I1f729fbd86801e7649c92087583dca539e2d1aa1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1263
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>