Add health check cron job to readyset-server
Adds a health check script to readyset-server that is run every five
minutes by cron. The script tries to establish a tcp connection with
noria server (using a default timeout of 30 minutes). If the connection
fails, the script notifies the ASG that the current EC2 instance is
unhealthy.

The script uses a temp file to prevent multiple health checks from
running at once. The health check task will not run if the temp file was
modified within an interval equal to the default timeout. Otherwise, it
will update the temp file timestamp and begin the health check. When the
health check completes, it will modify the timestamp again such that a
new health check can begin immediately.

If the health check task dies while executing, health checks will resume
again after the timestamp on the temp file is older than the default
timeout.

The script does not run health checks until 30 minutes after the server
is brought up.  This is to give time for the readyset-server service to
start running.  This can be improved by starting the health checks only
after ensure-ebs-volume reports that it is finished.

Change-Id: I24c5e45d2b1c51edf768a94a6be563dcbd243b05
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1265
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>
Tested-by: Buildkite CI