persistent_state: sync+WAL for writing replication offset
We've seen a persistent issue which we can't quite pin down where writes
to the metadata of a persistent state (specifically setting the
replication offset to a non-None value) are persistently getting lost
between server restarts - we've tracked this down to some non-obvious
rocksdb behavior where not calling set_sync(true) on the write options
seems to never actually write to disk. This is super scary! But we know
about it now - and the checkpoint in our system for having fully
snapshot a table is the replication offset getting written, so in this
case we can just set_sync to true if the replication offset exists, even
if snapshot mode is enabled.

Change-Id: I7327d093e917c8cceca4eef488c38b201500decf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1269
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>