fault tolerance: Correctly clear channel coordinator on new leader.
On leader restart, the dataflow graph is reconstructed from an
empty state by repeating the recipe. As a result, we clear the
existing domains on each of the workers. However, we did not
clear the channel coordinator which may store connections
for a NodeIndex to domains that no longer exist and will result
in failing to send messages.

CL add's clearing the channel coordinator.

Change-Id: I2e9bfb370a5f1804df1d5ad88b13426ff3cfe330
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1284
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>