persistent_state: Switch to JSON for PersistentMeta
bincode is great for super low space-overhead,
super-high-efficiency (de)serialization, but in part *because* of the
lack of value tagging it uses to be so low space-overhead, is remarkably
difficult to make properly backwards compatible for struct types.
Specifically, in our case, we want to add a new field to the
PersistentMeta struct, but doing so in a backwards compatible way has
proven prohibitively difficult (or even impossible - we never really got
further than the attempt made in
https://gerrit.readyset.name/c/readyset/+/1254, which is failing).

None of the reasons we'd choose bincode really apply to meta, however,
and JSON is way easier to make backwards compatible (just add
`#[serde(default)]` to new fields!), plus it has the extra advantage of
being human-readable for debugging purposes. So, this commit switches
the serialization and deserialization of the PersistentMeta to rocksdb
to use serde_json.

For no-manual-intervention backwards compatibility with existing
databases, this also makes deserialization of metadata from rocksdb
fallible - if deserialization fails, we just log the error and return
the default PersistentMeta value, which includes a None replication
offset so the table would just be re-snapshot. Then, when this is
detected during creation of a PersistentState, we go through all column
families in the db that are missing from the meta and drop them,
effectively clearing the database. This will result in correct behavior,
though it will potentially lock the upstream database for the snapshot
if this ever happens.

Change-Id: I84e9ff47a3268aeb4bd2ec50b55bca3ed1ddb11d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1288
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>