replicators: Set offset on all tables on log rotation
When we get a `ReplicationAction::LogPosition` replication action, which
happens when mysql rotates the binlogs, persist that offset to all
tables via the table `set_replication_offset` RPC in addition to storing
it on the schema offset. This is nice generally as it keeps replication
offsets close to each other, but also especially matters if we lose a
table and have to re-snapshot, as that means we'll have less catching up
to do between the old tables and the new replication offset
post-snapshot.

Change-Id: I7b2a88c5c8b3029db1126f06a47231f73959908a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1292
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>