ENG-859: Attempts to get base table indices from non-existant inc.
A migration adds a new recipe to the head of a linked list of
recipes. When that occurs we either clone or steal the
SqlIncorporator from the previous head. However, when performing
the `activate` step of a migration, we attempt to retrieve base
table node indices from the prior recipe. In cases where the
SqlIncorporator is stolen, this leads to throwing an error since
the incorporator doesn't exist.

This CL updates `activate` to instead use the current incorporator.
This is valid because the current incorporator is either the
incorporator from the prior recipe (if it was taken), or it is a
clone of the prior recipe.

Change-Id: I59670d24f53c773a0abb0c3cf1ccab2cc45b30ae
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1303
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>