fault tolerance: Re-retrieve views after a failed query.
Each backend in Noria caches the set of tables: `inputs` and
views `outputs`. A view maps to a single server+node pair, and
once it is cached in a backend, there is no mechanism for removing
it. This is problematic as views can point to dead servers or the
incorrect NodeIndex.

This CL adds re-retrieving views on the subsequent query after a
query fails on a view. This is implemented as a hashset with the
queries that need to re-retrieve their views.

An alternate implementation removes the view from Outputs immediately
on failure, however, optimizations require us hold borrows to a View
through the entire response path.

Change-Id: I909bbd0336ace182ff28632e95906a322def1dc7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1302
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>