clustertest: Views to failed servers never recover.
The adapter caches views for clients but does not remove views that
have become invalid: either refer to server that have failed or
node index's that are no longer valid. In both these cases,
ReadySet throws errors, but we do not clear the view from the cache.

This results in the adapter going to fallback for any query to a
view that has failed. This adds a clustertest that shows this
behavior.

Refs: ENG-615
Change-Id: I061f52802d60186882deac2f3c7e898bc9f179ba
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1301
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI