test: Install recipe twice with a slightly modified DDL.
This adds an #[ignore] integration test that checks the behavior
of the system when the DDL passed to `install_recipe` varies
slightly. This can occur when we snapshot as a result of a
failure. DDL can differ if it was received via the binlog
vs. as a result of snapshotting and as a result, fail the
noria-server.

Refs: ENG-863
Change-Id: I24331260cee2c349e53f3c60f48ab2578bf5e5da
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1312
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>