ops/cfn: mysql cloudformation stack - use SSM and deprecate use of dbUrl
The purpose of this CL is to move us away from plain-text database URLs
and instead, leverage parameter store to retrieve the database password
for the backend database we're providing caching for.

A new script has been added to external-base to allow for automatic
concatenation and exporting of the database URL as an env var. This will
allow us to centralize management of the script and the action itself.

Please note that there are significant limitations with native
secure-ssm calls in CloudFormation, as only certain resource types and
certain properties on them support calls to SSM via CloudFormation
directly.

Users should encrypt their parameters with KMS, and provide the ARN of
the KMS key so we can grant our IAM roles access to consume such
secrets.

Added documentation for the new required SSM dbPassword parameter and
respective KMS encryption key. I staged up a default value and encrypted
with the AWS managed SSM KMS key, so it should be readily available to
anyone trying to deploy via our usual means.

Change-Id: Ica50131f918aa7cc041c52b9231b3ae477e6b8cf
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1305
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>
Tested-by: Buildkite CI