noria: Moved the dataflow state into its own module and made it thread-safe.
TL;DR
A new structure was created to hold the dataflow state information.
For the sake of having small CLs, this just introduces the skeleton for the
rest of the ENG-805 work.
See `state.rs`, which has most of the changes.
------------------------------------------------------

The whole Noria dataflow state was scattered as fields on the `Leader`
structure, but now they are in their own structure (`DataflowState`) in their
own module.
Along with the new `DataflowState`, a couple other structures were created as
well, to act as wrappers and allow thread-safe access to the dataflow state.
Right now, and until all the old logic from the `Leader` gets moved to its new
place in the `DataflowState`, the added structures are unused.
Once all the methods are migrated (in subsequent CLs), we'll swap the current
dataflow state in the `Leader` for a `DataflowStateHandle` (see below).
Here's a summary of the new structures that are introduced (please refer to the
rustdocs for a full explanation of how they work and their
invariants/guarantees):

-----------------------
|`DataflowStateHandle`|
-----------------------
This is the structure that replaces all the fields in the `Leader`.
The main idea is that it holds a `DataflowStateReader` (see below) behind a
`tokio::sync::RwLock` (which is safe to send across threads) which readers can
take; and a `tokio::sync::Mutex` which is used to allow only one writer at a
time.

-----------------------
|`DataflowStateReader`|
-----------------------
This thin wrapper allows the threads to have read-only access to its underlying
`DataflowState` (by implementing only `Deref`).
Can only be created inside the `state` module, by the `DataflowStateHandle`.
The only way to modify the underlying state, is by swapping it with a new one,
and only the `DataflowStateHandle` can do that (internally). Users are not able
to modify the state.

-----------------------
|`DataflowStateWriter`|
-----------------------
This wrapper allows the threads to have write and read access to its underlying
`DataflowState` (by implementing `Deref` and `DerefMut`).
Can only be created inside the `state` module, by the `DataflowStateHandle`.
The `DataflowStateHandle` will try to lock its `Mutex` when creating a new
`DataflowStateWriter`, which will hold its guard.
So only one instance of `DataflowStateWriter` can exist at any given time (per
`DataflowStateHandle` instance, but there should only be one).

Refs: ENG-805
Change-Id: I5c90d436135e11c18a022716a712a19abc5a1017
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1138
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>