benchmarks: write throughput benchmark.
This adds a write throughput benchmark parameterized on a customer
schema. It generates writes uniformly at random across the set of
tables in the schema.

Sample usage:
  ./benchmarks write-benchmark
    --mysql-conn-str <mysql://>
    --schema <schema file>

This is accomplished by using a single generator per table made
thread-safe by a Mutex. It uses the same logic as the existing
data generator, see `parallel_load`.

Change-Id: Ie8e0dd7b4c4e0ab62d07adcaad0dfa919fd457a8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1315
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>