installer: Init, with prompt for deployment name
Add a skeleton of what will become a fully automated, interactive
ReadySet installer script. As of right now this currently only includes
management of deployments. At start-up we either:

* Prompt for the name of a new deployment to create, if none exist
  already,
* Confirm to continue with an existing deployment, if only one exists,
  or
* Select from multiple existing deployments, if more than one exists.

The deployments themselves are saved in an either a user-supplied
directory, or an automatically-created `readyset` subdirectory of the
`$XDG_STATE_HOME` directory (tbd for what we do on MacOS).

For prompts, this is using the dialoguer crate[0], which provides really
nice, interactive terminal prompts including text inputs, boolean
confirmations, selects, etc.

Change-Id: I9d23fcd887cbab0e41473737d9d63b2ac287d955
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1318
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>