Ensure MigrationHandler always uses live conn
We encountered a failure state where if the connection between the
MigrationHandler and the upstream database is ever dropped, we fail to
continue reconciling queries.

The failure state would be hit the moment a connection between the
MigrationHandler (previously named QueryReconciler) and upstream
database (mysql/postgres) is dropped. The MigrationHandler relies on an
implementer of the UpstreamDatabase trait. This allows us to support
both mysql and postgres easily. In the implementations of these traits,
we changed them to use normal connections rather than pools so that we
can properly support prepared statements to fallback.

This works fine for the normal adapter code path. If a conn to upstream
is dropped, we drop the conn to the client, causing them to retry their
connection, and we in turn establish a fresh conn to the backing
database.

In the case of the MigrationHandler, we expect that connection to be
good forever, which is basically an impossible assumption. In a real
system that conn will drop - it's just a matter of time. Once it does,
the MigrationHandler will never mark new queries as allowed, causing all
queries from that point forward to go to fallback.

We also have no way to recover from it right now other than force
restart the adapter instances.

This commit creates a way for us to reset the connection to upstream,
and a way to detect if we got back a connection closed error while
trying to prepare a query in the migration handler. This should ensure
that we reset our connection when it has dropped.

Change-Id: Ie4f1d60e3354f37e2712de068a5207b452415c09
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1320
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>