PersistentIndex: Use indirect indexing for secondary indices
This CL changes the way data is stored in the base tables.
Previosuly data was copied for each new index added, with this
change we instead only store the pointer to the main data index.

Althought this change requires two lookups for anything that is
not in the primary index, in practice the performance penalty is
not great, and is offset by previous optimizations to the
PersistentIndex, and also by the fact that we now use much less
storage and memory, which should be a net boon for performance.

In terms of storage, for a table that consumed 5575 MiB of storage,
previously adding two secondary indices resulted in it growing to
17710 MiB, or more than 200% overhead. With this change, adding the
same indices results in 7237 MiB, or just 30% overhead.

Reindexing is now also more than twice as fast, since less data
needs to be copied around.

Change-Id: Id9ebac1998081b704fbfa15a359f6696ec118d5d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1326
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>