benchmarks: Split up deployment parameters and benchmark parameters.
This splits up the parameters that a user needs to pass to get the
benchmark harness to work, from the set of parameters that a user
requires to run a benchmark in a certain way.

This allows us to have benchmark specifications that are independent
of the deployment. See changes in between local_cache_hit.yaml to
cache_hit.yaml. This comes at the cost of being able to check some
user parameters in detail. I will add spitting out exactly the
benchmark that is running to the user in a followup.

Sample usage:
./benchmarks
  --target-conn-str mysql://127.0.0.1:3307
  --setup-conn-str: mysql://127.0.0.1:3306/test
  --from-file cache_hit.yaml

In a separate follow-up I will add specifying the deployments
specification from yaml. This can be used to speed up local dev as
these parameters often do not change and can be read from the same
static file each time.

Change-Id: I8d4c7a38cfe1015f28c6dd32f595b855f26000ba
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1322
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>