mir: Remove Rc loop by replacing ancestor Rc with Weak.
An Rc loop, when two Rc objects refer to eachother, is a way of
indirectly leaking memory in Rust. Objects that are in Rc loops
always have a valid reference and are never dropped.

This is particularly a problem in noria-logictest where we create
tens of thousands of mir nodes over the course of the nightly
pipeline that are persisted for the entire duration due to Rc
loops.

This CL replaces the mir graph's ancestor references with weak
pointers. When appropriate we upgrade these pointers to strong
pointers to make modifications. Since the mir graph is stored
from both the leafs and the base nodes in MirQuery, this *should*
be safe and not lead to the graph being dropped before intended.

Fixes: ENG-871
Change-Id: I00b45fdfcf499b6604b1edc3ab214549c61342e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1329
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>