installer: Prompt for AWS credentials profile
Step 1 of the installer is to check the AWS credentials we're using to
make sure they have all the permissions we require. AWS can store
multiple credentials profiles in the ~/.aws/credentials directory, so
before we do that we need to prompt for the profile to use - then store
it in the saved Deployment.

This starts (but doesn't go all the way) with generalizing the pattern
of configuring something, then saving it to the deployment, then using
that saved value to pick up where we left off next time. Specifically in
the case of this commit, this adds a custom pair of serialization and
deserialization functions, usable with the `#[serde(with = "...")]`
field attribute, which serialize an Option but use the string "unset"
instead of `null` as the value for None, to allow explicitly
differentiating between null values and unset values.

Fixes: ENG-874
Change-Id: I36701ffc7c359916d0c32000f5dbe56e242653a8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1331
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>