installer: Prompt for AWS VPC
After configuring the AWS credentials and region, prompt to either
create a new AWS VPC or install into an existing one - in the latter
case, provide the user with a list of VPCs in their configured region
and allow them to select which VPC they'd like to use. As with all other
user config, this is then serialized to the Deployment to allow them to
pick up where they left off.

Change-Id: I377031166abc385ef5d85836d292e29a9cfb4843
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1334
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>