cfn: Update Bastion Stack AMI
Using Ubuntu 20.04 in the AWS provided Bastion ctack currently installs
a Python 2 version of the CloudFormation scripts. The Python 2 version
is now entirely broke as one of the dependencies, which is also not
locked in any way, has been updated to only work on Python 3.6 and
later.

Updating the AMI to Amazon Linux allows the testing to continue for now
until we build a bastion host or our own or merge this with the monitor
instance.

This causes no issues for IRL as IRL is not deploying the Bastion stack
we are using here and it is just for testing.

Change-Id: Iab2a82b53d45ad01fcd76c99b4129db67d2624e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1339
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>
Reviewed-by: Marcus Bastian <marcus@readyset.io>
Tested-by: Buildkite CI