installer: Ensure VPC exists, and has 3 subnets
After we've either loaded the VPC ID to deploy into from the existing
deployment or the user has selected a VPC to deploy into, and if the
user *hasn't* elected to have us create a new VPC for them, validate
that the VPC has 3 subnets in different availability zones. If not, tell
the user, then prompt to automatically create subnets in the remaining
AZs for them. If the user selects "yes", this just bails with a TODO
error message (to be implemented in a later commit), or if they select
"no", this exits with an error message telling them to create subnets in
the missing AZs and then re-run the installer.

On the side, there's some generalization of logging, plus a utility in
`launchpad` for formatting English lists with commas and
conjunctions (like "and" or "or") in the right place.

Change-Id: I497431f04517bb268a7953e0a2e2dbabb264d3ff
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1341
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Michael Cronce <mc@readyset.io>