installer: Create missing subnets
If the user gives us permission to, create subnets in the selected
VPC to bring us up to three subnets in three AZs.

In order to do this, we need to figure out the CIDR blocks to use for
those subnets, which is surprisingly annoying given any number of
networks could be taken up by subnets already in the VPC. Essentially,
we use the `iprange` crate to figure out the set of network space *not*
already in the VPC, then divide that up amongst the number of subnets we
need to create, up to the next power of two for the CIDR bitmask. There
are unit tests for this, plus a property test that checks that we either
return an error indicating there wasn't enough room in the VPC, or
return the correct number of non-overlapping subnets.

Change-Id: I8e9de891be03b015f59b11d694117bc5e04677db
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1345
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>