installer: Prompt for RDS db instance
After successfully validating the VPC, prompt for whether to deploy in
front of an existing RDS db instance or create a new one as part of the
installer process. In the former case, prompt for an RDS database
selecting from a list of DB instances in the already-selected VPC. In
the latter case, prompt the user for whether to use MySQL or PostgreSQL
as their engine. In either case, store a MaybeExisting with the RDS DB
ID and the selected engine in the Deployment

Change-Id: Idfb42762e787b40c83fa6bfaba357de76b4843f4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1349
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI