Add a snapshot benchmark
The snapshot benchmark measures the time it takes
to perform a snapshot of a given database.

It reports the time to finish the snapshot and the
amount of disk space used.

Optionally it can also extend the schema after the
snapshot and report how long the migration to the
new schema took, and how much disk space is used
following the migration.

Example of output:

Snapshot time:     32 s
Tables replicated: 10
Disk space used:   246.71 MiB
Migration time:    4 s
Disk space used:   349.21 MiB

Change-Id: I59e15c3ec05ca57ba71824bbd2b5d96ad31860d4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1353
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>