Fixed failing types::timestamp_tz_datetime test.
The test itself wasn't failing, but the proptest framework was not being able to
instantiate the test data (of type `DateTime<FixedOffset>`), since the
generation strategy was wrong.
The generation strategy was creating a `FixedOffset` by calling the method
`FixedOffset::west(secs)`, where secs was an integer and -86_400 <= secs <=
86_400; but `FixedOffset::west(secs)` specifically requires that those bounds
are -86_400 < secs < 86_400.
The issue was fixed by generating seconds in the range -86_399..86_399.

This commit also adds a more verbose way of failing for this particular case, so
in the future we know if the `FixedOffset` generation failed, and what value was
used.

Change-Id: I3005aa77a921a449ddef30414543357a9464cbd3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1354
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>