installer: Validate and auto-fix MySQL RDS config
If the user wants to install in front of an existing MySQL RDS database,
validate that that database's parameter group has binlog_format=ROW and
that the backup_retention_period is set to a value > 0. If this isn't
the case, give the user the option to automatically fix the
configuration. If they accept, fix the configuration by creating a new
parameter group if necessary with all the same parameters as the
existing RDS database except binlog_format, which is set to ROW, and
updating the DB to set that as its parameter group and set the backup
retention period to 1 day.

There's a *lot* of complexity here in the interaction with the AWS API,
in no small part because the AWS SDK for rust irritatingly has all
fields as Options, but this is under a pretty aggressive deadline so
there's not a ton I can do to reduce that at the moment - in the future,
I would like to refactor this to split out some of the complexity.

Change-Id: Ic6004f492e46eef3c6d63d5a683ee8251f126b56
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1355
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>