installer: Use directories crate for state dir
Rather than rolling our own XDG directory manipulation, which currently
isn't cross-platform, use the `directories` crate, which *is*
cross-platform, to figure out where to store serialized deployments
for the installer. This uses the `data_dir` instead of the `state_dir`
which we were previously using, since the latter doesn't exist on MacOS
or windows while the former does.

Change-Id: Ida592a933549cc3efa4ace88c1eda3eb29ae1e64
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1358
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>