Store rewritten queries in query_status_cache
Currently queries are not rewritten before being placed in the
query_status_cache. This CL adds an extra call to
rewrite::process_query() in the adapter just for the purpose of reading
from and manipulating the query status cache. It also adds a map of
prepared statement id to rewritten query to avoid the rewrite step when
we execute the query.

This change is necessary for two reasons:
- When the OutputsSynchronizer polls the list of migrated queries, it
  will read expressions from Recipe which have been rewritten. These
expressions must correspond to the statements used to access and
manipulate the query status cache.
- The query_status_cache could grow by the number of queries if
  customers interpolate query parameters.

Change-Id: I85f5747197d9b5b744868b7cfdbc1171d3429a06
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1360
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>