Support for dynamic failure injection in noria-server with failpoints.
With the fail-rs crate, we are able to set dynamically configurable
failpoints in our code. This enables us to annotate pieces of code
so that we can trigger specific failure behaviors. This CL builds
upon this functionality to allow setting noria-server wide failpoint
behavior through a controller RPC `failpoint`.

For manual testing, `failpoint` in the `noria/tools` binary allows
setting failpoints on any noria-server. The clustertest framework
is also expanded to include `set_failpoint` on server's to
allow programatic testing of injected failures.

The nightly and per-commit pipelines are also updated to build
noria-server with failure_injection for clustertests.

See docs/src/failure_injection.md for more details on how to
inject failures with failpoints.

Change-Id: I38b6d24195a4ae2a40c4b2efe35406a2fd2ac2d8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1403
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>