ops/tailscale: complete sandbox rollout of new tailscale implementation
The purpose of this CL is to take what we did in eu-west-1 in a previous
CL, and scale that out to all regions within the Sandbox AWS account.

Some minor refactoring to the module has been included to help make
things a bit more dynamic, but also supportive of changes in direction
wrt using CMKs vs AWS managed keys for encrypting the Secrets Manager
objects.

One extra improvement included here is to set the hostname of the node
to something we understand. This will give us more clarity on what we're
looking at in the TS admin console, which is always nice, in addition to
seeing a clear name in Tailscale on client machines, etc.

After deploying the changes included here, we're able to see 5 new
Tailscale subnet router nodes in the Tailscale administrative console.

We should we able to go in and clean up the old EC2 instances from the
aforementioned console as well @Harley! Please feel free to do
additional testing on this :)

Change-Id: Ia5f0aae6d32011cf629e5c8e77a03f91b3301157
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1368
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI