installer: Allow using a preexisting key pair
Before prompting for the name of a key pair to create, give the user the
option to use an *existing* ssh key pair, to avoid every invocation of
the installer from having to create a brand new ssh key pair.

I'm doing this now mostly because it significantly eases testing the
installer, since I'm running it over and over again.

Change-Id: I43e01ccd30f27c044ad83b7cc7522a5399d8b5bc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1374
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>