installer: Create VPC if necessary
If the user has selected not to deploy with a preexisting VPC, create
one for them by deploying the official AWS quickstart VPC stack, using
the same settings as our various super-templates. Then, wait for the
stack to enter the CREATE_COMPLETE status, and once it does, take the
VPCID and the three subnet IDs from the stack outputs and store them in
the Deployment.

Change-Id: Icc503c758548043496813a257a41e9d3862717f2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1375
Tested-by: Buildkite CI
Reviewed-by: Michael Cronce <mc@readyset.io>