ops/tailscale: build environment rollout of new tailscale infra
This CL is a continuation of the work detailed in the submitted CL:
https://gerrit.readyset.name/c/readyset/+/1368

Effectively, the purpose of this CL is to get the new Tailscale subnet
routers deployed into all regions within the context of the build AWS
account.

All subnet routers have been deployed via Terraform and are checking
into the master control plane. Routing has also been enabled for these
new nodes in the Tailscale control plane, so we should be good to go.

Change-Id: Idb8e3bc4abff82f95a9e7a6bf67632aa2d1a1713
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1376
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI