noria/data: Fix symmetry on Float == Double
The way this was implemented would convert the f32 to an f64 on Double ==
Float comparisons, but the f64 to an f32 on Float == Double comparisons
- the latter case is lossy in the case that the f64 is bigger than would
fit in an f32, causing the equivalence relation to not be symmetric.
This is unlikeely to happen, but was eventually caught in CI by a
failing proptest, which is committed to proptest-regressions in this
commit. yay proptest!

Over in the dataflow expression tests, this broke a test that was
comparing an f32 DataType to an f64 DataType and relying on this
asymmetric conversion to work; specifically:

    >> 4.123_f32 as f64
    4.123000144958496
    >> 4.123_f64
    4.123

Obviously these two values aren't going to be equal if we're upcasting
things to f64s when comparing f32s and f64s. This is actually fine,
since float conversion being lossy is a pretty well established fact,
and it's much more important that the equivalence relation given by the
Eq impl for datatype is an actual equivalence relation.

Change-Id: Ie9cbd0624fff77cc4e2244340dd5b28a2a2b120f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1388
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>